+++ src/training/steps/market_analysis/hybrid_nas_tas_regime/comprehensive_demo.py
@@ -19,7 +19,7 @@ from src.utils.tprint import (
 )
 
 # Import unified regime detection system
-from src.utils.ml_common.nas_tas_unified import (
+from src.utils.nas_tas import (
     UnifiedRegimeDetector, UnifiedRegimeConfig, RegimeDetectionMethod,
     OptimizationStrategy, EconomicEvaluationMode,
     PerformanceOptimizer, get_performance_optimizer,
