+++ src/training/steps/market_analysis/hybrid_nas_tas_regime/shared_utils/ML_COMMON_INTEGRATION_README.md
@@ -279,7 +279,7 @@ def analyze_market_regimes(self, market_data, timestamps=None, enable_multi_time
    - ✅ Updated initialization, search, and evaluation methods
    - ✅ Updated metadata to reflect shared utilities usage
 
-3. **Unified Regime Detector** (`src/utils/ml_common/nas_tas_unified/unified_regime_detector.py`)
+3. **Unified Regime Detector** (`src/utils/nas_tas/unified_regime_detector.py`)
    - ✅ Unified TAS and NAS regime detection with shared utilities
    - ✅ Performance optimization and real-time monitoring
    - ✅ Comprehensive benchmarking and configuration options
