+++ src/training/steps/market_analysis/hybrid_nas_tas_regime/unified_regime_integration.py
@@ -21,7 +21,7 @@ from src.utils.tprint import (
 
 # Import unified regime system
 try:
-    from src.utils.ml_common.nas_tas_unified import (
+    from src.utils.nas_tas import (
         UnifiedRegimeDetector, UnifiedRegimeConfig, UnifiedRegimeResult,
         RegimeDetectionMethod, OptimizationStrategy, EconomicEvaluationMode
     )
