+++ src/training/steps/market_analysis/nas_regime/core/perfect_nas_regime_detector.py
@@ -22,7 +22,7 @@ from src.utils.tprint import (
 
 # Import unified utilities
 try:
-    from src.utils.ml_common.nas_tas_unified import (
+    from src.utils.nas_tas import (
         UnifiedRegimeDetector, UnifiedRegimeConfig, UnifiedRegimeResult,
         RegimeDetectionMethod
     )
