+++ src/training/steps/market_analysis/nas_regime/core/perfect_nas_regime_detector_legacy.py
@@ -24,7 +24,7 @@ from src.utils.tprint import (tprint, tprint_debug, tprint_info, tprint_warning,
 
 # Import unified utilities
 try:
-    from src.utils.ml_common.nas_tas_unified import (
+    from src.utils.nas_tas import (
         UnifiedRegimeDetector, UnifiedRegimeConfig, UnifiedRegimeResult,
         RegimeDetectionMethod
     )
