+++ src/training/steps/market_analysis/tas_regime/core/tas_regime_detector.py
@@ -21,7 +21,7 @@ from src.utils.tprint import (
 
 # Import unified utilities
 try:
-    from src.utils.ml_common.nas_tas_unified import (
+    from src.utils.nas_tas import (
         UnifiedRegimeDetector, UnifiedRegimeConfig, UnifiedRegimeResult,
         RegimeDetectionMethod
     )
